feat: metadata keys collection (#2384)
* First draft of schema and dto for metadata keys

* Added template for module, controller and service

* fix lint and import url

* added controller & dto

* some lint fix

* added migration script

* improve migration script

* add casl-ability for metadata-key and add missing fields for dto and schema

* fix wrong swagger  response type for attachment v4 endpoint

* added findAll endpoint with casl ability

* remove create, update, delete abilities for metadataKeys

* remove corresponding action enum from the above commit

* fix ApiResponse type for attachment and metadatakeys controller

* update default frontend config for more compact table content

* add access logic

* integrate metadataKeys service with dataset

* unit test fix

* fix unit test

* updated documentation

* cleanup unused import

* fix typo

* create metadatakeys for dataset, instrument, proposal and samples

* fix unit test

* add unit tests for metadatakeys v4

* refactor metadatakeys service unit test to make it  readable

* add api test

* fix api test

* test1

* test2

* test

* fix metadatakey api test

---------

+++ src/datasets/datasets.service.ts
@@ -59,6 +59,10 @@ import {
   DatasetLookupKeysEnum,
 } from "./types/dataset-lookup";
 import { ProposalsService } from "src/proposals/proposals.service";
+import {
+  MetadataKeysService,
+  MetadataSourceDoc,
+} from "src/metadata-keys/metadatakeys.service";
 
 @Injectable({ scope: Scope.REQUEST })
 export class DatasetsService {
@@ -67,15 +71,30 @@ export class DatasetsService {
     private configService: ConfigService,
     @InjectModel(DatasetClass.name)
     private datasetModel: Model<DatasetDocument>,
-    private datasetsAccessService: DatasetsAccessService,
-    @Inject(ElasticSearchService)
-    private elasticSearchService: ElasticSearchService,
     @Inject(REQUEST) private request: Request,
+
+    private datasetsAccessService: DatasetsAccessService,
+    private elasticSearchService: ElasticSearchService,
+    private metadataKeysService: MetadataKeysService,
     private proposalService: ProposalsService,
   ) {
     if (this.elasticSearchService.connected) {
       this.ESClient = this.elasticSearchService;
     }
+  }
+
+  private createMetadataKeysInstance(
+    doc: UpdateQuery<DatasetDocument>,
+  ): MetadataSourceDoc {
+    const source: MetadataSourceDoc = {
+      sourceType: "dataset",
+      sourceId: doc.pid,
+      ownerGroup: doc.owner,
+      accessGroups: doc.accessGroups || [],
+      isPublished: doc.isPublished || false,
+      metadata: doc.scientificMetadata ?? {},
+    };
+    return source;
   }
 
   addLookupFields(
@@ -164,17 +183,22 @@ export class DatasetsService {
       // insert created and updated fields
       addCreatedByFields(createDatasetDto, username),
     );
-    if (this.ESClient && createdDataset) {
-      await this.ESClient.updateInsertDocument(createdDataset.toObject());
-    }
 
     const savedDataset = await createdDataset.save();
+
+    if (this.ESClient && createdDataset) {
+      await this.ESClient.updateInsertDocument(savedDataset.toObject());
+    }
 
     if (savedDataset.proposalIds && savedDataset.proposalIds.length > 0) {
       await this.proposalService.incrementNumberOfDatasets(
         savedDataset.proposalIds,
       );
     }
+
+    this.metadataKeysService.insertManyFromSource(
+      this.createMetadataKeysInstance(savedDataset),
+    );
 
     return savedDataset;
   }
@@ -384,12 +408,12 @@ export class DatasetsService {
   async findByIdAndReplace(
     id: string,
     updateDatasetDto: UpdateDatasetDto,
-  ): Promise<DatasetDocument> {
+  ): Promise<DatasetDocument | null> {
     const username = (this.request.user as JWTUser).username;
     const existingDataset = await this.datasetModel.findOne({ pid: id }).exec();
 
     if (!existingDataset) {
-      throw new NotFoundException();
+      throw new NotFoundException(`Dataset #${id} not found`);
     }
     // TODO: This might need a discussion.
     // NOTE: _id, pid and some other fields should not be touched in any case.
@@ -414,9 +438,13 @@ export class DatasetsService {
       throw new NotFoundException(`Dataset #${id} not found`);
     }
 
-    if (this.ESClient && updatedDataset) {
+    if (this.ESClient) {
       await this.ESClient.updateInsertDocument(updatedDataset.toObject());
     }
+
+    await this.metadataKeysService.replaceManyFromSource(
+      this.createMetadataKeysInstance(updatedDataset),
+    );
     // we were able to find the dataset and update it
     return updatedDataset;
   }
@@ -451,27 +479,50 @@ export class DatasetsService {
       )
       .exec();
 
-    if (this.ESClient && patchedDataset) {
+    // check if we were able to find the dataset and update it
+    if (!patchedDataset) {
+      throw new NotFoundException(`Dataset #${id} not found`);
+    }
+
+    if (this.ESClient) {
       await this.ESClient.updateInsertDocument(patchedDataset.toObject());
     }
+
+    await this.metadataKeysService.replaceManyFromSource(
+      this.createMetadataKeysInstance(patchedDataset),
+    );
     // we were able to find the dataset and update it
     return patchedDataset;
   }
 
   // DELETE dataset
   async findByIdAndDelete(id: string): Promise<DatasetDocument | null> {
+    const deletedDataset = await this.datasetModel
+      .findOneAndDelete({
+        pid: id,
+      })
+      .exec();
+
+    if (!deletedDataset) {
+      throw new NotFoundException(`Dataset #${id} not found`);
+    }
+
     if (this.ESClient) {
       await this.ESClient.deleteDocument(id);
     }
-    const deletedDataset = await this.datasetModel.findOneAndDelete({
-      pid: id,
-    });
 
     if (deletedDataset?.proposalIds && deletedDataset.proposalIds.length > 0) {
       await this.proposalService.decrementNumberOfDatasets(
         deletedDataset.proposalIds,
       );
     }
+
+    // delete metadata keys associated with this dataset
+    await this.metadataKeysService.deleteMany({
+      sourceId: id,
+      sourceType: "dataset",
+    });
+
     return deletedDataset;
   }
   // GET datasets without _id which is used for elastic search data synchronization
